[ci skip] Update Syncs - Rename to Managers
