+++ Includes/Enums/CardEnums.h
@@ -3729,12 +3729,6 @@ BETTER_ENUM(ZodiacYear, int, INVALID = -1, PRE_STANDARD = 0, KRAKEN = 1,
 
 BETTER_ENUM(Zone, int, INVALID = 0, PLAY = 1, DECK = 2, HAND = 3, GRAVEYARD = 4,
             REMOVEDFROMGAME = 5, SETASIDE = 6, SECRET = 7)
-
-BETTER_ENUM(TaskID, int, INVALID = 0, TASK_TUPLE = 1, REQUIRE = 2,
-            USER_SETTING = 3, SWAP = 4, SHUFFLE = 5, DRAW = 6, MODIFY_MANA = 7,
-            MODIFY_HEALTH = 8, BRIEF = 9, SELECT_MENU = 10, SELECT_CARD = 11,
-            SELECT_TARGET = 12, MULLIGAN = 13, SUMMON_MINION = 14, COMBAT = 15,
-            PASS_TASKMETA = 16, GAME_END = 17)
 }
 
 #endif
